Refactor: Rename views to align with CRUD operations
